src/parser.py
import aiohttp
from bs4 import BeautifulSoup
import re
import json
from typing import Optional

class RecipeParser:
    def __init__(self):
        self.session: Optional[aiohttp.ClientSession] = None
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36',
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
            'Accept-Language': 'ru-RU,ru;q=0.9,en-US;q=0.8,en;q=0.7',
        }
    
    async def _get_session(self) -> aiohttp.ClientSession:
        if not self.session:
            self.session = aiohttp.ClientSession(headers=self.headers)
        return self.session
    
    async def parse_recipe(self, url: str) -> str:
        try:
            session = await self._get_session()
            
            async with session.get(url, timeout=30) as response:
                response.raise_for_status()
                html = await response.text()
            
            # Парсим HTML
            soup = BeautifulSoup(html, 'lxml')
            
            # Удаляем скрипты и стили
            for tag in soup(['script', 'style', 'nav', 'header', 'footer', 'aside']):
                tag.decompose()
            
            # Извлекаем текст
            text = soup.get_text(separator='\n', strip=True)
            
            # Очищаем текст
            lines = [line.strip() for line in text.split('\n') if line.strip()]
            text = '\n'.join(lines)
            
            # Ограничиваем длину
            if len(text) > 50000:
                text = text[:50000] + "..."
            
            print(f"✅ Извлечено {len(text)} символов")
            return text
            
        except Exception as e:
            print(f"❌ Ошибка парсинга: {e}")
            # Возвращаем тестовый рецепт если сайт недоступен
            return self._get_test_recipe()
    
    def _get_test_recipe(self) -> str:
        """Тестовый рецепт если сайт не отвечает"""
        return """
        Борщ классический
        
        Ингредиенты:
        - Говядина - 500 г
        - Свекла - 2 шт
        - Капуста - 300 г
        - Картофель - 3 шт
        - Морковь - 1 шт
        - Лук - 1 шт
        - Томатная паста - 2 ст.л.
        - Чеснок - 3 зубчика
        - Соль, перец по вкусу
        
        Приготовление:
        1. Сварить бульон из говядины 1 час
        2. Нарезать свеклу соломкой, потушить
        3. Обжарить лук и морковь с томатной пастой
        4. Добавить картофель в бульон
        5. Добавить капусту
        6. Добавить зажарку и свеклу
        7. Варить 15 минут
        8. Добавить чеснок и специи
        """
    
    async def close(self):
        if self.session:
            await self.session.close()
            self.session = None
