requirements.txt
# requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
beautifulsoup4==4.12.2
readability-lxml==0.8.1
lxml==5.1.0
python-dotenv==1.0.0